# specs/002-todo-fullstack-webapp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Full-Stack Web Application API
  description: |
    RESTful API for the Todo Full-Stack Web Application (Phase II).
    All task endpoints require JWT authentication via Better Auth.
  version: 1.0.0
  contact:
    name: Hackathon II Team

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todo-app.example.com
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Tasks
    description: Task CRUD operations

security:
  - BearerAuth: []

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns the health status of the API
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: |
        Retrieve all tasks for the authenticated user.
        Supports filtering by completion status.
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - name: status
          in: query
          description: Filter by completion status
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Tasks
      summary: Create a task
      description: Create a new task for the authenticated user
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/{user_id}/tasks/{task_id}:
    get:
      tags:
        - Tasks
      summary: Get a task
      description: Retrieve a specific task by ID
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Tasks
      summary: Update a task
      description: Update task title and/or description
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Tasks
      summary: Delete a task
      description: Permanently delete a task
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle task completion
      description: Toggle the completed status of a task
      operationId: toggleTaskComplete
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
        - $ref: '#/components/parameters/TaskIdPath'
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth

  parameters:
    UserIdPath:
      name: user_id
      in: path
      required: true
      description: The authenticated user's ID
      schema:
        type: string
        example: "user_abc123"

    TaskIdPath:
      name: task_id
      in: path
      required: true
      description: The task's unique identifier
      schema:
        type: integer
        minimum: 1
        example: 42

  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 42
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Optional task description
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp (UTC)
          example: "2025-12-29T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: "2025-12-29T10:30:00Z"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (required)
          example: "Buy groceries"
        description:
          type: string
          maxLength: 2000
          description: Optional task description
          example: "Milk, eggs, bread"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated task title
          example: "Buy groceries and fruits"
        description:
          type: string
          maxLength: 2000
          nullable: true
          description: Updated task description
          example: "Milk, eggs, bread, apples"

    TaskResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Task'

    TaskListResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'

    ErrorResponse:
      type: object
      required:
        - detail
        - status_code
        - request_id
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Task not found"
        status_code:
          type: integer
          description: HTTP status code
          example: 404
        request_id:
          type: string
          description: Unique request identifier for debugging
          example: "req_abc123xyz"
        code:
          type: string
          description: Machine-readable error code
          example: "NOT_FOUND"

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Title is required"
            status_code: 400
            request_id: "req_abc123"
            code: "VALIDATION_ERROR"

    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Invalid or expired token"
            status_code: 401
            request_id: "req_abc123"
            code: "INVALID_TOKEN"

    Forbidden:
      description: User not authorized to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "You can only access your own data"
            status_code: 403
            request_id: "req_abc123"
            code: "FORBIDDEN"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"
            status_code: 404
            request_id: "req_abc123"
            code: "NOT_FOUND"